+++ packages/expo-image-picker/ios/ImagePickerModule.swift
@@ -167,7 +167,7 @@ public class ImagePickerModule: Module, OnMediaPickingResultHandler {
   // MARK: - OnMediaPickingResultHandler
 
   func didCancelPicking() {
-    self.currentPickingContext?.promise.resolve(["cancelled": true])
+    self.currentPickingContext?.promise.resolve(ImagePickerResponse(assets: nil, canceled: true))
     self.currentPickingContext = nil
   }
 
@@ -191,7 +191,7 @@ public class ImagePickerModule: Module, OnMediaPickingResultHandler {
     mediaHandler.handleMultipleMedia(selection) { result -> Void in
       switch result {
       case .failure(let error): return promise.reject(error)
-      case .success(let response): return promise.resolve(response.dictionary)
+      case .success(let response): return promise.resolve(response)
       }
     }
   }
@@ -214,7 +214,7 @@ public class ImagePickerModule: Module, OnMediaPickingResultHandler {
     mediaHandler.handleMedia(mediaInfo) { result -> Void in
       switch result {
       case .failure(let error): return promise.reject(error)
-      case .success(let response): return promise.resolve(response.dictionary)
+      case .success(let response): return promise.resolve(response)
       }
     }
   }