+++ packages/expo-image-picker/ios/ImagePickerResponse.swift
@@ -1,51 +1,15 @@
 // Copyright 2022-present 650 Industries. All rights reserved.
 
+// swiftlint:disable redundant_optional_initialization
+// Unfortunately, property wrappers must be initialized in those records, otherwise the memberwise initializer
+// would require `Field<FieldType?>` as an argument instead of `FieldType?`.
+// TODO: (@tsapeta) Figure out if we can fix that
+
 import ExpoModulesCore
 
 internal typealias ImagePickerResult = Result<ImagePickerResponse, Exception>
 
-internal typealias SelectedMediaResult = Result<SelectedMediaInfo, Exception>
-
-/**
- General protocol representing a picker response
- */
-internal protocol ImagePickerResponse {
-  var dictionary: [String: Any] { get }
-}
-
-/**
- Represents a picker response, when multiple selection is disabled
- */
-internal enum ImagePickerSingleResponse: ImagePickerResponse {
-  case image(ImageInfo)
-  case video(VideoInfo)
-
-  var dictionary: [String: Any] {
-    var result: [String: Any] = [:]
-    switch self {
-    case .video(let videoInfo):
-      result = videoInfo.dictionary
-    case .image(let imageImage):
-      result = imageImage.dictionary
-    }
-    result["cancelled"] = false
-    return result
-  }
-}
-
-/**
- Represents a picker response, when multiple selection is enabled
- */
-internal struct ImagePickerMultipleResponse: ImagePickerResponse {
-  let results: [SelectedMediaInfo]
-
-  var dictionary: [String: Any] {
-    [
-      "cancelled": false,
-      "selected": results.map { $0.dictionary }
-    ]
-  }
-}
+internal typealias SelectedMediaResult = Result<AssetInfo, Exception>
 
 /**
  Convenience alias, a dictionary representing EXIF data
@@ -53,69 +17,25 @@ internal struct ImagePickerMultipleResponse: ImagePickerResponse {
 internal typealias ExifInfo = [String: Any]
 
 /**
- General protocol representing a single selected asset
+ Represents a picker response.
  */
-internal protocol SelectedMediaInfo {
-  var dictionary: [String: Any] { get }
+internal struct ImagePickerResponse: Record {
+  @Field var assets: [AssetInfo]? = nil
+  @Field var canceled: Bool = true
 }
 
 /**
- Represents a single selected image
+ Represents a single asset (image or video).
  */
-internal struct ImageInfo: SelectedMediaInfo {
-  let assetId: String?
-  let type: String = "image"
-  let uri: String
-  let width: Double
-  let height: Double
-  let fileName: String?
-  let fileSize: Int?
-  let base64: String?
-  let exif: ExifInfo?
-
-  var dictionary: [String: Any] {
-    var result: [String: Any] = [
-      "type": type,
-      "uri": uri,
-      "assetId": assetId,
-      "width": width,
-      "height": height,
-      "fileName": fileName,
-      "fileSize": fileSize
-    ]
-    if base64 != nil {
-      result["base64"] = base64
-    }
-    if exif != nil {
-      result["exif"] = exif
-    }
-    return result
-  }
-}
-
-/**
- Represents a single selected video
- */
-internal struct VideoInfo: SelectedMediaInfo {
-  let assetId: String?
-  let type: String = "video"
-  let uri: String
-  let width: Double
-  let height: Double
-  let fileName: String?
-  let fileSize: Int?
-  let duration: Double
-
-  var dictionary: [String: Any] {
-    [
-      "type": type,
-      "uri": uri,
-      "assetId": assetId,
-      "width": width,
-      "height": height,
-      "fileName": fileName,
-      "fileSize": fileSize,
-      "duration": duration
-    ]
-  }
+internal struct AssetInfo: Record {
+  @Field var assetId: String? = nil
+  @Field var type: String = "image"
+  @Field var uri: String = ""
+  @Field var width: Double = 0
+  @Field var height: Double = 0
+  @Field var fileName: String? = nil
+  @Field var fileSize: Int? = nil
+  @Field var base64: String? = nil
+  @Field var exif: ExifInfo? = nil
+  @Field var duration: Double? = nil
 }